junit: print Spike trace to stdout
Spike was run by bash - that may cause 'zombie' Spike
processes on server (bash process is killed by test, but
Spike may be alive). It was made because of the absence of
"--write-log-to-file" option of Spike. Now Spike prints
it's log to stdout - for simplicity. The problem will be
solved more carefully in future.

Signed-off-by: Sergey Smolov <smolov@ispras.ru>